Implement functions
Add functions to the lexer and parser
Add function and function call AST
Add labeling on functions and calls
Add function calls to the bytecode
Fix some newly emerging bugs in parsing parenthesis

The code generation will pass over the list of functions to declare them
before it goes into the labelling stage. This means you can declare
functions in any order and everything should just work. The labelling
environment now maintains a list of function name/number mappings as well
as the local variable mappings. In order to support functions, the
environment dictionary was wrapped in a class that supports a lookup
hierarchy.

The lookup hierarchy doesn't actually work because symbols in the outer
scope will be hidden by symbols in the inner scope with the same number.
TODO: To fix this, we need to add a mechanism like Python's LOAD_GLOBAL
instruction.