Add some lexer tests and fix exposed problems
Most of the parsing test failures were caused by numbers. Now we handle
exponential (1.2e+2, etc) notation generated by Python. Parsing tests will
still reject negative numbers.

Also fixed an issue where the lexer would treat a sequence like "iffoo"
as the sequence of tokens "if" and "foo", which is clearly not the
expected behavior. Keywords now have to be followed by a non-identifier
character in order to be lexed as such.